refactor: 初期画面のUI描写をviewに移動
